+++ packages/patternfly-4/react-core/src/components/Modal/Modal.test.js
@@ -1,14 +1,13 @@
 import Modal from './Modal';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { shallow } from 'enzyme';
+import { mount, shallow } from 'enzyme';
 import { KEY_CODES } from '../../helpers/constants';
+import { css } from '../../../../react-styles/dist/js';
+import styles from '@patternfly/patternfly/components/Backdrop/backdrop.css';
 
-jest.spyOn(ReactDOM, 'createPortal');
 jest.spyOn(document, 'createElement');
 jest.spyOn(document, 'addEventListener');
-
-ReactDOM.createPortal.mockImplementation(v => v);
 
 const props = {
   title: 'Modal',
@@ -43,5 +42,15 @@ test('modal does not call onClose for esc key if it is not open', () => {
 test('Each modal is given a new id', () => {
   const first = shallow(<Modal {...props} />);
   const second = shallow(<Modal {...props} />);
-  expect(first.props().id).not.toBe(second.props().id);
+  expect(first.instance().id).not.toBe(second.instance().id);
+});
+
+test('modal removes body backdropOpen class when removed', () => {
+  const TestRemoval = testProps => (testProps.display ? <Modal {...props} isOpen /> : <p>Not displayed</p>);
+  const view = mount(<TestRemoval display />);
+  view.update();
+  expect(document.body.className).toContain(css(styles.backdropOpen));
+  view.setProps({ display: false });
+  view.update();
+  expect(document.body.className).not.toContain(css(styles.backdropOpen));
 });